RottenPotatoes Enhancement #1: Sorting the list of all movies
You will enhance RottenPotatoes in the following ways:

On the "All Movies" page, the column headings for "Movie Title" and
"Release Date" for a movie should be clickable links. Clicking one of
them should cause the list to be reloaded but sorted in ascending order
on that column. Namely,
clicking the "Movie Title" column should list the movies alphabetically
by title (for movies whose names begin with non-letters, the sort order
should match the behavior of String#<=>), and
clicking the "Release Date" column heading should redisplay the list of
movies with the earliest-released movies first.
When the listing page is redisplayed with sorting-on-a-column enabled,
the column header that was selected for sorting should appear with a
yellow background, as shown below. You should do this by setting
controller variables that are used to conditionally set the CSS class
of the appropriate table heading to hilite, and pasting this simple CSS
snippet into RottenPotatoes's app/assets/stylesheets/application.css
file.