ROTTENPOTATOES ENHANCEMENT #2: FILTER THE LIST OF MOVIES
Specific Requirements

Make sure that you don't break the sorted-column functionality you
added in part 1B! That is, sorting by column headers should still work,
and if the user then clicks the "Movie Title" column header to sort by
movie title, the displayed results should both be sorted and be limited
by the Ratings checkboxes.
If the user checks (say) 'G' and 'PG' and then redisplays the list, the
checkboxes that were used to filter the output should appear checked
when the list is redisplayed. This will require you to modify the
checkbox form slightly from the version we provided above.
The first time the user visits the page, all checkboxes should be
checked by default (so the user will see all movies). For now, ignore
the case when the user unchecks all checkboxes—you will get to this in
the next part.