ROTTENPOTATOES ENHANCEMENT #3: REMEMBER THE SETTINGS
Specific Requirements:
If the user explicitly includes new sorting/filtering settings in
params[], the values stored in the session should not override them. On
the contrary, the new settings should be remembered in the session.
If a user unchecks all checkboxes, use the settings stored in the
session[] hash (it doesn't make sense for a user to uncheck all the
boxes).